Added code to output dates in which particular courses were played.
The weather data needs this information, in addition to the geocoded locations
of the courses, to determine the lat/lon and particular days on which to
pull historical weather data.

The get_course_active_dates.py file does nothing but call a shotlink module
method; the tests test the module method.