Updated Andrew Smith's code to use shot data that doesn't require previous processing.
The last change updated the geocoding code to use a file that was itself generated
using geocoded locations. This could be fine because that file does exist, now that
we've already generated it, but it does mean that a user that just has access to the
stock ShotLink data couldn't run it. Now the code uses just stock ShotLink data.
I also made a few other tweaks to make it run, including updating the output filename
to not use an undefined constant.